main.js split into data.js, util.js, main.js
